Change inference.py to a module file
